+++ ntex-rt/src/arbiter.rs
@@ -5,6 +5,7 @@ use std::{cell::RefCell, collections::HashMap, fmt, future::Future, pin::Pin, th
 
 use async_channel::{Receiver, Sender, unbounded};
 
+use crate::Handle;
 use crate::system::{FnExec, Id, System, SystemCommand};
 
 thread_local!(
@@ -29,6 +30,7 @@ pub struct Arbiter {
     id: usize,
     pub(crate) sys_id: usize,
     name: Arc<String>,
+    hnd: Handle,
     sender: Sender<ArbiterCommand>,
     thread_handle: Option<thread::JoinHandle<()>>,
 }
@@ -108,19 +110,23 @@ impl Arbiter {
 
         let name = name2.clone();
         let sys_id = sys.id();
+        let (arb_hnd_tx, arb_hnd_rx) = oneshot::channel();
 
         let handle = builder
             .spawn(move || {
                 log::info!("Starting {name2:?} arbiter");
-
-                let arb = Arbiter::with_sender(sys_id.0, id, name2, arb_tx);
 
                 let (stop, stop_rx) = oneshot::channel();
                 STORAGE.with(|cell| cell.borrow_mut().clear());
 
                 System::set_current(sys);
 
-                config.block_on(async move {
+                crate::driver::block_on(config.runner.as_ref(), async move {
+                    let arb = Arbiter::with_sender(sys_id.0, id, name2, arb_tx);
+                    arb_hnd_tx
+                        .send(arb.hnd.clone())
+                        .expect("Controller thread has gone");
+
                     // start arbiter controller
                     crate::spawn(
                         ArbiterController {
@@ -149,8 +155,11 @@ impl Arbiter {
                 panic!("Cannot spawn an arbiter's thread {:?}: {:?}", &name, err)
             });
 
+        let hnd = arb_hnd_rx.recv().expect("Could not start new arbiter");
+
         Arbiter {
             id,
+            hnd,
             name,
             sys_id: sys_id.0,
             sender: arb_tx2,
@@ -169,6 +178,7 @@ impl Arbiter {
             sys_id,
             name,
             sender,
+            hnd: Handle::current(),
             thread_handle: None,
         }
     }
@@ -183,6 +193,13 @@ impl Arbiter {
         self.name.as_ref()
     }
 
+    /// Handle to a runtime
+    pub fn handle(&self) -> &Handle {
+        &self.hnd
+    }
+
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::spawn()`")]
     /// Send a future to the Arbiter's thread, and spawn it.
     pub fn spawn<F>(&self, future: F)
     where
@@ -193,12 +210,13 @@ impl Arbiter {
             .try_send(ArbiterCommand::Execute(Box::pin(future)));
     }
 
-    #[rustfmt::skip]
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::Handle::spawn()`")]
     /// Send a function to the Arbiter's thread and spawns it's resulting future.
     /// This can be used to spawn non-send futures on the arbiter thread.
     pub fn spawn_with<F, R, O>(
         &self,
-        f: F
+        f: F,
     ) -> impl Future<Output = Result<O, oneshot::RecvError>> + Send + 'static
     where
         F: FnOnce() -> R + Send + 'static,
@@ -216,11 +234,15 @@ impl Arbiter {
         rx
     }
 
-    #[rustfmt::skip]
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::Handle::spawn()`")]
     /// Send a function to the Arbiter's thread. This function will be executed asynchronously.
     /// A future is created, and when resolved will contain the result of the function sent
     /// to the Arbiters thread.
-    pub fn exec<F, R>(&self, f: F) -> impl Future<Output = Result<R, oneshot::RecvError>> + Send + 'static
+    pub fn exec<F, R>(
+        &self,
+        f: F,
+    ) -> impl Future<Output = Result<R, oneshot::RecvError>> + Send + 'static
     where
         F: FnOnce() -> R + Send + 'static,
         R: Send + 'static,
@@ -234,6 +256,8 @@ impl Arbiter {
         rx
     }
 
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::Handle::spawn()`")]
     /// Send a function to the Arbiter's thread, and execute it. Any result from the function
     /// is discarded.
     pub fn exec_fn<F>(&self, f: F)
@@ -247,17 +271,22 @@ impl Arbiter {
             })));
     }
 
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::set_item()`")]
     /// Set item to current arbiter's storage
     pub fn set_item<T: 'static>(item: T) {
-        STORAGE
-            .with(move |cell| cell.borrow_mut().insert(TypeId::of::<T>(), Box::new(item)));
+        set_item(item);
     }
 
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::get_item()`")]
     /// Check if arbiter storage contains item
     pub fn contains_item<T: 'static>() -> bool {
         STORAGE.with(move |cell| cell.borrow().get(&TypeId::of::<T>()).is_some())
     }
 
+    #[doc(hidden)]
+    #[deprecated(since = "3.8.0", note = "use `ntex_rt::get_item()`")]
     /// Get a reference to a type previously inserted on this arbiter's storage
     ///
     /// # Panics
@@ -352,4 +381,23 @@ impl ArbiterController {
             }
         }
     }
+}
+
+/// Set item to current runtime's storage
+pub fn set_item<T: 'static>(item: T) {
+    STORAGE.with(move |cell| cell.borrow_mut().insert(TypeId::of::<T>(), Box::new(item)));
+}
+
+/// Get a reference to a type previously inserted on this runtime's storage
+pub fn get_item<T: 'static, F, R>(f: F) -> R
+where
+    F: FnOnce(Option<&mut T>) -> R,
+{
+    STORAGE.with(move |cell| {
+        let mut st = cell.borrow_mut();
+        let item = st
+            .get_mut(&TypeId::of::<T>())
+            .and_then(|boxed| boxed.downcast_mut());
+        f(item)
+    })
 }